fix bugs in sending email and change quantity to decimal

+++ FishShop.API/Contracts/CreateOrderItem.cs
@@ -3,5 +3,5 @@
 public record CreateOrderItem
 {
     public int ProductSizeId { get; init; }
-    public int Quantity { get; init; }
+    public decimal Quantity { get; init; }
 }
